+++ src/services/deviationsFeeds/PriceTriggerFilter.ts
@@ -6,7 +6,7 @@ import {DeviationFeed, FilterResult, PriceData} from '../../types/DeviationFeeds
 @injectable()
 export class PriceTriggerFilter {
   apply(deviationFeed: DeviationFeed, leaf: Leaf, priceData: PriceData): FilterResult {
-    const newPrice = this.currentPrice(leaf);
+    const newPrice = this.currentPrice(leaf, deviationFeed.precision);
     const priceDiff = this.abs(newPrice - priceData.price);
     const percent = Number((priceDiff * 10000n) / priceData.price) / 100;
 
@@ -14,7 +14,7 @@ export class PriceTriggerFilter {
       return {result: false, msg: `${leaf.label}: flat price`};
     }
 
-    const trigger = this.triggerAmount(priceData.price, deviationFeed.trigger);
+    const trigger = this.triggerAmount(priceData.price, deviationFeed.trigger, deviationFeed.precision);
     const triggerFired = priceDiff >= trigger;
     let msg = '';
 
@@ -27,13 +27,15 @@ export class PriceTriggerFilter {
     return {result: triggerFired, msg};
   }
 
-  protected currentPrice(leaf: Leaf): bigint {
+  protected currentPrice(leaf: Leaf, precision: number): bigint {
     // price in leaf is 18 decimals, price for deviations is 8 decimals => we need to divide by 1e10
-    return BigInt(leaf.valueBytes) / 10n ** 10n;
+    return BigInt(leaf.valueBytes) / 10n ** (18n - BigInt(precision));
   }
 
-  protected triggerAmount(price: bigint, trigger: number): bigint {
-    const percent = BigInt(Math.trunc(trigger * 1e8));
+  protected triggerAmount(price: bigint, trigger: number, precision: number): bigint {
+    // first we multiply by 1e8 to not overflow Number, 1e8 it is enough precision for trigger
+    // then we normalize percent to desired precision
+    const percent = BigInt(Math.trunc(trigger * 1e8)) * 10n ** (18n - BigInt(precision));
     // price * % / 100
     return (price * percent) / 100_00000000n;
   }